[ME] Share cooperative frame admission and separate Cubemap UI lifetime
