+++ src/MaterialEditor.Base/MaterialEditor.Base.projitems
@@ -83,7 +83,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)PluginBase.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.PropertyOrganizer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\Assets\UI.CubemapImportCoordinator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialEditorCubemapImportCoordinator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialFrameWorkScheduler.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\Assets\UI.CubemapImportRunner.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\Assets\MaterialEditorAssetWorkflow.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.RepositoryContract.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\UI.SessionState.cs" />
